photo/frame/requirements.txt
micropython-picographics
micropython-jpegdec
